+++ backend/src/main/java/edonymyeon/backend/post/application/PostService.java
@@ -1,9 +1,5 @@
 package edonymyeon.backend.post.application;
 
-import static edonymyeon.backend.global.exception.ExceptionInformation.MEMBER_ID_NOT_FOUND;
-import static edonymyeon.backend.global.exception.ExceptionInformation.POST_ID_NOT_FOUND;
-import static edonymyeon.backend.global.exception.ExceptionInformation.POST_MEMBER_NOT_SAME;
-
 import edonymyeon.backend.global.exception.EdonymyeonException;
 import edonymyeon.backend.image.ImageFileUploader;
 import edonymyeon.backend.image.domain.Domain;
@@ -152,98 +148,5 @@ public class PostService {
                 imageFileUploader.uploadFiles(request.newImages()));
         post.updateImages(updatedPostImageInfos);
         postImageInfoRepository.saveAll(updatedPostImageInfos.getPostImageInfos());
-    }
-
-    public List<GeneralPostInfoResponse> findPostsByPagingCondition(
-            final GeneralFindingCondition generalFindingCondition) {
-        PageRequest pageRequest = convertConditionToPageRequest(generalFindingCondition);
-        final Slice<Post> foundPosts = postRepository.findAllBy(pageRequest);
-        return foundPosts
-                .map(post -> GeneralPostInfoResponse.of(post, domain.getDomain()))
-                .toList();
-    }
-
-    private static PageRequest convertConditionToPageRequest(final GeneralFindingCondition generalFindingCondition) {
-        try {
-            return PageRequest.of(
-                    generalFindingCondition.getPage(),
-                    generalFindingCondition.getSize(),
-                    switch (generalFindingCondition.getSortDirection()) {
-                        case ASC -> Direction.ASC;
-                        case DESC -> Direction.DESC;
-                    },
-                    generalFindingCondition.getSortBy().getName()
-            );
-        } catch (IllegalArgumentException e) {
-            throw new EdonymyeonException(ExceptionInformation.POST_INVALID_PAGINATION_CONDITION);
-        }
-    }
-
-    @Transactional
-    public SpecificPostInfoResponse findSpecificPost(final Long postId, final MemberId memberId) {
-        final Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new EdonymyeonException(POST_ID_NOT_FOUND));
-
-        final Optional<Member> member = memberRepository.findById(memberId.id());
-        post.updateView(member.orElse(null));
-
-        final ReactionCountResponse reactionCountResponse = new ReactionCountResponse(
-                post.getViewCount(),
-                0, // TODO: 댓글 수 기능 구현 필요
-                0 // TODO: 스크랩 기능 구현 필요
-        );
-        final AllThumbsInPostResponse allThumbsInPost = thumbsService.findAllThumbsInPost(postId);
-        final WriterDetailResponse writerDetailResponse = getWriterResponse(post.getMember());
-
-        if (member.isEmpty()) {
-            return SpecificPostInfoResponse.of(
-                    post,
-                    allThumbsInPost,
-                    writerDetailResponse,
-                    reactionCountResponse,
-                    domain.getDomain()
-            );
-        }
-
-        final ThumbsStatusInPostResponse thumbsStatusInPost = thumbsService.findThumbsStatusInPost(memberId, postId);
-
-        return SpecificPostInfoResponse.of(
-                post,
-                allThumbsInPost,
-                writerDetailResponse,
-                reactionCountResponse,
-                thumbsStatusInPost,
-                member.get(),
-                domain.getDomain()
-        );
-    }
-
-    private WriterDetailResponse getWriterResponse(final Member member) {
-        if (Objects.isNull(member.getProfileImageInfo())) {
-            return new WriterDetailResponse(
-                    member.getId(),
-                    member.getNickname(),
-                    null
-            );
-        }
-        return new WriterDetailResponse(
-                member.getId(),
-                member.getNickname(),
-                domain.getDomain() + member.getProfileImageInfo().getStoreName()
-        );
-    }
-
-    public GeneralPostsResponse searchPosts(final String searchWord,
-                                            final GeneralFindingCondition generalFindingCondition) {
-        final Specification<Post> searchResults = PostSpecification.searchBy(searchWord);
-        final PageRequest pageRequest = convertConditionToPageRequest(generalFindingCondition);
-
-        final Slice<Post> foundPosts = postRepository.findAll(searchResults, pageRequest);
-
-        List<GeneralPostInfoResponse> posts = foundPosts.stream()
-                .map(post -> GeneralPostInfoResponse.of(post, domain.getDomain()))
-                .toList();
-
-        return new GeneralPostsResponse(posts);
     }
 }
