+++ backend/src/test/java/edonymyeon/backend/post/application/PostFindingSpecificPostTest.java
@@ -38,7 +38,8 @@ public class PostFindingSpecificPostTest {
     private final MemberRepository memberRepository;
     private final ProfileImageInfoRepository profileImageInfoRepository;
     private final MemberTestSupport memberTestSupport;
-    private final PostService postService;
+    private final PostReadService postReadService;
+
     private MemberId memberId;
     private MemberId member2Id;
     private Long postId;
@@ -52,7 +53,7 @@ public class PostFindingSpecificPostTest {
 
     @Test
     void 게시글아이디가_주어지면_게시글의_상세정보를_알려준다() {
-        final var postInfoResponse = postService.findSpecificPost(postId, memberId);
+        final var postInfoResponse = postReadService.findSpecificPost(postId, memberId);
 
         assertAll(
                 () -> assertThat(postInfoResponse.title()).isEqualTo("Summer Breeze"),
@@ -73,7 +74,7 @@ public class PostFindingSpecificPostTest {
 
     @Test
     void 로그인_되어있지_않으면_조회는_가능하되_추천여부와_스크랩여부와_작성자여부는_모두_false이다() {
-        final var postInfoResponse = postService.findSpecificPost(postId, new AnonymousMemberId());
+        final var postInfoResponse = postReadService.findSpecificPost(postId, new AnonymousMemberId());
 
         assertThat(postInfoResponse.isUp()).isFalse();
         assertThat(postInfoResponse.isDown()).isFalse();
@@ -83,14 +84,14 @@ public class PostFindingSpecificPostTest {
 
     @Test
     void 작성자_본인이_본인_게시글을_보는경우_isWriter값이_true이다() {
-        final var postInfoResponse = postService.findSpecificPost(postId, memberId);
+        final var postInfoResponse = postReadService.findSpecificPost(postId, memberId);
 
         assertThat(postInfoResponse.isWriter()).isTrue();
     }
 
     @Test
     void 타인의_게시글을_보는경우_isWriter값이_false이다() {
-        final var postInfoResponse = postService.findSpecificPost(postId, member2Id);
+        final var postInfoResponse = postReadService.findSpecificPost(postId, member2Id);
 
         assertThat(postInfoResponse.isWriter()).isFalse();
     }
